Rename Video component
- Move liveness intro to steps
